probando repositorioio

+++ src/main/java/com/example/edelec/repositories/CarreraRepository.java
@@ -2,8 +2,10 @@ package com.example.edelec.repositories;
 
 import com.example.edelec.entitys.Carrera;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarreraRepository extends JpaRepository<Carrera, Integer> {
+public interface CarreraRepository extends JpaRepository<Carrera, Long> {
+
 }
